Refactoring; move Slack interaction endpoint to API resource instead of web route
