+++ servicex/resources/users/delete_user.py
